backend/requirements.txt
# Web Framework and Extensions
Flask==3.0.2
flask-cors==4.0.0
flask-jwt-extended==4.6.0
python-dotenv==1.0.1
Werkzeug==3.0.1

# MongoDB
pymongo==4.6.1
